Conflicts:
	dataobj/einstellungen.cc
	dataobj/fahrplan.cc
	dings/baum.cc
	dings/gebaeude.cc
	dings/leitung2.cc
	gui/citylist_frame_t.cc
	gui/convoi_info_t.cc
	gui/depot_frame.cc
	gui/fahrplan_gui.cc
	gui/map_frame.cc
	gui/money_frame.cc
	gui/schedule_list.cc
	gui/stadt_info.cc
	player/simplay.cc
	simcity.cc
	simconvoi.cc
	simhalt.cc
	simtools.cc

+++ dataobj/dingliste.cc
@@ -1022,7 +1022,7 @@ void dingliste_t::rdwr(karte_t *welt, loadsave_t *file, koord3d current_pos)
 		ding_t *save[256];
 		sint32 max_object_index = 0;
 		for(  uint16 i=0;  i<top;  i++  ) {
-			ding_t *d=bei(i);
+			ding_t *d = bei((uint8)i);
 			if(d->get_typ()==ding_t::way
 				// do not save smoke
 				||  d->get_typ()==ding_t::raucher
